refactor(sandbox): extract detachMount to shared module
